Don't store parent in Widget class
I haven't used it anywhere, and honestly I'd like to keep it that way, because I feel it could lead to bad practice.
With this out of the way, I can make children public, which can be convenient.
In this commit I'm also making events public, because why not.
Also I'm adding a check for Display.NONE to compute_layout, in case you have multiple root widgets.