Add Display.FLOAT
With this option, parent position is ignored, and the widget just uses the position from its style instead.
I don't know what happens if you have float widgets and regular ones with the same parent, and I don't know how size will behave, but neither of those are really supported use cases.
This is just for movable windows on a screen.