Add Window and Screen widgets
The window can be moved and resized (the latter not implemented yet), and has a header and frame.
It requires three events, which it subscribes to internal callbacks.
The header is subscribed to the mouse press and release events; it's a pretty clever solution @JonasAlmaas came up with.
The mouse move callback has some math to keep the window on the screen.

Speaking of which, the screen covers the 3D view, is invisible by default, has windows as children, and manages which window is focused.
When you click inside the area of a window, it is moved to the front. It uses an approach that's not as elegant as I'd like, but it works.
Only the focused window actually gets to handle events.
In the future we might want to make an exception for scroll events, which at least on Windows, can go to unfocused windows too.