Overhaul event system again
Now, you can have events that don't care about modifier keys.
This comes at a performance cost, probably, but it's not like you'll notice.
One event can trigger multiple subscriptions on a widget, but will still only be sent to its children once.
ModalState is just constructed when you use it, not updated over time, because I decided to simplify it.
EventCopy is back to being ModalEvent, because it's not really a copy of a Blender event anymore, and has no class method for doing that.
Screen window focusing and escape handling are done with subscriptions now.
I made some other changes but they're not too important.