Fix float
Instead of removing the sum of sizes of floating children, just make them return 0 in the first place.
Added to width as well (it was only height previously).
Using inside position as origin for floating children now, instead of content position.
The code is a little ugly but it works.