Adjust code structure
Layout now takes the modal state, which it uses in compute_width/height to make it so flexing widgets with no parent take the size of the 3D view.
I moved the logic from Widget.compute_layout and Widget..render to compute_layout and render_widget, so those are just single function calls now.
Render takes the modal state too now instead of the context, because it's our thing and I like using it.
Screen makes use of the new layout behavior by setting its width and height to flex instead of overwriting the compute_layout method.
I adjusted the example operator to work with the new code also.