Overhaul everything
Oh boy, there's a lot.
Text is now just a string, its style just lives inside Style; that means no more TextStyle.
The remove methods for Image and Font have a little safety now, oh and you can create a Font without a path, which will just use ID 0.
Style color is background_color now, text color is foreground_color.
Styles can be added together (the one to the right of the operand overwrites things from the one to the left).
The computed style and layout of a widget are private now, as are its parent and children.
The user has access to 3 styles, which all start as None (all the parameters on a new Style are also None btw).
The regular style is added to the default style, then the hovered style is added if we are hovering, then the pressed style is added if we are pressed by any mouse button (this seems to be how it works in CSS too).
The event system is completely overhauled, to behave more like Qt, because I think it makes more sense.
Widget comes with a bunch of abstract on_* methods now, which are called when it receives different types of events.
With this big commit out of the way, I've got some cleaning up to do.