Add selected style to widget
Widget._select is a bool that can be set by the user, which in this example is done when the item is clicked (well it's on release but you get what I mean).
I renamed hovered to just hover, and added a bool for active, which is just a bool cast of pressed (so it's True if any mouse buttons are pressed on the widget).
Widget.on_event now only checks the mouse position on mouse move events, and just uses the hover variable for other events.
The styles are applied in this order: `DEFAULT_STYLE` `style` `style_select` `style_hover` `style_active`.