Give widget a list of styles with criteria for when to use them
The styles are applied in the order you added them.
Each style can have a set of criteria that determines whether it should be used.
Hover and active are set for you by Widget.on_event, select and focus you'll have to set yourself.
I can add classes for Selectable and Focusable in the future perhaps, or make these options on Widget.
I added a nice hover effect for selected items in the example.