Implement weight for relative sizes
Size is a different kind of class now. Instead of using a float for absolute sizes, you use Size.absolute() with your value. You use Size.relative() with a 0-1 value (you can use more on the main axis; I wouldn't recommend it on the cross axis).
Size.children/texture() are shorthands for Size(type=Size.Type.CHILDREN/TEXTURE).
I changed the order of if blocks in the layout functions a bit, aside from updating it to work with the new names and implementing weights for relative sizes.
I updated the example operator as well, though it doesn't actually showcase weights at the moment (I did test though; it works fine).
Also I'm using `is` instead of `==` for enum values in Size.Type now because apparently that's the recommended way (though the other one is fine too); will probably do the same for other enum checks in the code later.