Make images internal
By adding a dot at the start of the name, the images are hidden in the list.
This means the user can't see their thumbnails, which prevents crash if the operator removes images while the user is in the list.
I'm prepending `.bwl.` rather than just a dot, to improve the odds of a unique name.