Support mouse movement in scroll boxes
Now you can click and drag in scroll boxes and it'll move them.
It has to be in empty space though; the items take precedence.
If I allow you to drag on the items too, by not eating mouse press events on them, then the item under the cursor will eat the mouse release event, and the scroll box won't stop following the cursor.
One possible solution would be to return a boolean from the on_* method, to for example make the item only care about left click, and the scroll box only care about middle click.
I'll probably want to do this anyway actually. Or maybe get rid of the categories and just have a single on_event method, or make it so you're expected to overwrite the handle method.
Another would be to handle release events non-exclusively (return False from the handle method), and then in the item keep track of the parent's scroll, and only consider it clicked if the scroll didn't change more than 10 pixels or so.
In the taskbar on Windows, dragging only starts around 10 pixels away from your click, so it's the same idea but with swapped priority. Probably makes more sense that way, gives it some inertia.