Return if an event was handled from on_* methods
You're expected to return True only if the event was handled exclusively, otherwise False or None.
This will be useful when a widget listens for left click, while its parent listens for middle click, for example.
This change removes the need for abstract and is_abstract, so I'm just removing them.
The docstrings for the on_* methods don't mention that you're supposed to return True for handled events, due to lack of space. I'll try to come up with an elegant solution for this.
Oh and I got rid of the return at the end of the handle method, to show that None is also fine. Might change my mind about this later though.