Make window movable with middle click
Here's why I made it so you return whether an event was handled exclusively from the on_* methods, so a parent can listen to a different mouse button than their child, and both work.
I still need to come up with a way to keep the window inside the screen, but it's probably not that hard.