Clean up resources without 5 second delay
If you start the operator, close it, then start it again, the cleanup from the last instance could come in and delete resources used by the current instance.
I'm keeping the timer approach because it was necessary for headless operation (though I don't remember why exactly), but just reducing the delay from 5 seconds to as soon as Blender can run a the timer.